feature: compact.
Memory consumed for mapping 10 million keys is around ~410MB, which is
~29 bytes overhead per entry.

Enable the "compact" feature will try to shrink_to_fit() vector
allocations, which in turn can cost more CPU time (due to realloc).
But memory consumption reduces to ~310 MB, which is ~19 bytes overhead
per entry.

Note1: keys are 32 bit and values are 64 bit.

Note2: Memory measurement is based on the internal stats that accounts
for allocated memory, not the RSS memory accounted by the OS.